Dont compile test vectors into one big file
