+++ hosted/test-vectors/protocol/package.json
@@ -1,6 +1,6 @@
 {
   "scripts": {
-    "compile-vectors": "node scripts/compile-vectors.js"
+    "validate-vectors": "node validate-vectors.js"
   },
   "dependencies": {
     "ajv": "^8.0.1"
